Update README.md
Describe the description. As a Humananman the American Monkey, I naturally contradict myself, sometimes more brash & blatant than one would like to admit. While the statement in the previous paragraph refers to peace, it could just as easily be taken as the very antithesis to the tranquility it promotes. This inflicting of the will, however, does not refer to man-made-monkey's mode of manipulating their enviroment by any means necassary. No, no, quite contrare. Folllow the flow. This is mastery of the monkey mind, come along cosmonaut, Dulcinea didn't introduce the hammer to ours, now that, is the great wonder. Please, don't hesitiate.